requirements.txt
dlib==20.0.0
gaze-tracking==0.0.1
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.1
pillow==11.3.0
pyasn1==0.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
rsa==4.9.1
six==1.17.0
tzdata==2025.2
